requirements.txt
exifread
mutagen
piexif
Pillow
